server/requirements.txt
fastapi
uvicorn
asyncpg
sqlalchemy
python-dotenv
openai
chromadb
google-generativeai
